load float16 weight

+++ infer.py
@@ -19,7 +19,7 @@ if __name__ == '__main__':
     assert tokenizer.eos_token_id == 130005
     config.initializer_weight = False
     
-    pl_model = MyTransformer(config=config, model_args=model_args)
+    pl_model = MyTransformer(config=config, model_args=model_args, torch_dtype=torch.float16,)
 
     model = pl_model.get_llm_model()
     if not model.quantized:
